Refactor date range logic into shared utility and add tests
