+++ app/utils/__init__.py
@@ -28,3 +28,9 @@ __all__ = [
     "validate_required_fields",
     "format_error_response",
 ]
+
+from .date_ranges import get_date_range_from_params
+
+__all__ += [
+    "get_date_range_from_params",
+]
